added amplification, tested a couple, jankily fixed issues w/ first run
(amplifaction: duplicate item calls original's amplify to strengthen)

+++ Cookie Eater/src/ui/UIScoreCount.java
@@ -13,7 +13,7 @@ public class UIScoreCount extends UIElement{
 		parts.add(cash = new UIText(board,x+20,y+40,"",new Color(120,120,120,200),new Font("Arial",Font.BOLD,30)));
 		parts.add(cooks = new UIText(board,x,y,"",new Color(255,255,255,200),new Font("Arial",Font.BOLD,40)));
 	}
-	public void update(int ch, int ck, int nd) {
+	public void update(double ch, int ck, int nd) {
 		cash.setText("$"+ch);
 		cooks.setText(ck+"/"+nd);
 	}
